tools/docgen: improve code readability
In the docgen parser functions, reformat the code in order to avoid
lines too long, but only for few cases.

In genHtml, use the `tok` identifier when capturing the token from the
InlineSyntax and Shell nodes, since the scope is reduced and there is no
need to use long identifier.

Rename tokenizeAndPrintRaw to printZigCode and printShell to
printShellBlock, in order to have functions names consistent with
printSourceBlock.  Additionally, remove tokenizeAndPrint, since it is
not really necessary.