std.Build: model Module as a Step in the build graph
This representation eliminates the need for all the dependency
propagation logic which previously existed in `Module`. We need modules
to participate in the step graph, so that dependencies (e.g. a module
which includes generated header files) are propagated to the actual
steps which create them. Associating each `Module` with a dummy `Step`
means the build runner handles all of this for us! This change was
originally proposed in 537e280.

This commit also changes the `root_module` field of
`std.Build.Step.Compile` to be a `*Module` rather than a `Module`. This
is a breaking change, but an incredibly minor one (the full potential
extent of the breakage can be seen in the modified standalone test).
This change will be necessary for an upcoming improvement, so it was
convenient to make it here.