std: Add regression test for #18942
This change adds a regression test for
the issue reported in #18942 and fixed by PR #18945.

This test fails to compile on 0.12.0-dev.2750+5f9255829
(where this bug is present).
On versions where it compiles, it'll verify that bits for common flags
convert from C to Zig correctly.

The test will only compile on platforms where tc_flag_t is defined,
and has libc linked. These conditions are comptime-checked,
skipping the test if not set.