std.Build.Step.Compile: installHeader and friends depend on Compile step
Currently, `installHeader`, `installConfigHeader`, and friends all
add the installation step as a dependency of the root install step.
This means that if a library is installed with a custom top-level step,
then the headers are not installed.

To fix this, this commit changes these functions to instead have the
compile step depend on the header installation stpe.
This way, as long as the compile step for an artifact is consumed,
the headers will be installed.

Two sets of tests are included to verify this behavior:

- A unit test in std.Build.Step.Compile that verifies that dependencies
  are correctly connected.
- A standalone test that reproduces the issue reported in #17204, and
  verifies that the headers are installed even without the install step.

Resolves #17204