compiler: improve handling options with missing parameter
Currently the compiler frontend prints unfriendly error messages when
an option requires a parameter but it is not supplied by the command
line arguments.

Example:
  $ zig build-exe -x -fno-llvm main.zig
  error: language not recognized: '-fno-llvm'

Update the ArgsIterator.nextOrFatal method to report an error when the
argument is an option.

For compatibility, add a new nextMultiOrFatal method, that behaves in
the same way as the old nextOrFatal.  Currently it is only used for the
deprecated --mod option.

Example after this change:
  $ zig build-exe -x -fno-llvm main.zig
  error: expected parameter after -x

Additionally, improve the coding style used in ArgsIterator, removing
unnecessary use of @This and adding an empty line between methods.